requirements.txt
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.0
Pillow==10.2.0
moviepy==1.0.3
edge-tts==6.1.9
requests==2.31.0